(chore): move model factories to test/testUtils
