+++ examples/es6/basic/test/testUtils/factories/articles.js
@@ -1,0 +1,16 @@
+import faker from 'faker';
+import { Article } from '../../../src/models';
+
+const makeArticle = (overrides = {}) => {
+  return {
+    title: faker.lorem.word(),
+    body: faker.lorem.sentence(),
+    ...overrides,
+  };
+};
+
+const createArticle = (overrides = {}) => {
+  return Article.create(makeArticle(overrides));
+};
+
+export { makeArticle, createArticle };
